Support yearly periodicals.
Alma reponds differently when a periodical only has one issue each year.
Basically because the name of that issue is the year instead of a number during that year.
I chose the easy way out and simply called the issue '1'.

The best way would of course be to change ding_periodical to support yearly periodicals.